Bring in Dry and start building a pattern for mapping to rust
The code needed a way to test that we are mapping data correctly and I
felt that using dry-rb was a good way to go. This has allowed me to
start modeling the rust data types as structs instead of outputed text.
It also allows us to write tests against what the structs should output
as well as write tests against parsing the AST to the data objects.

Not all of the generator has been converted though.